wip: Java code working + experimental jni-rs
