+++ Cargo.toml
@@ -11,4 +11,5 @@ members = [
     "tests/foo-schema",
     "tests/foo-ffi",
     "tests/foo-bindings",
+    "tests/foo-java-ffi",
 ]
