+++ tests/foo-ffi/Cargo.toml
@@ -5,7 +5,7 @@ authors = ["Émile Grégoire <emile@automatak.com>"]
 edition = "2018"
 
 [lib]
-crate-type=["cdylib"]
+crate-type = ["rlib", "cdylib"]
 
 [dependencies]
