+++ tests/foo-java-ffi/Cargo.toml
@@ -1,11 +1,12 @@
 [package]
-name = "java-oo-bindgen"
+name = "foo-java-ffi"
 version = "0.1.0"
 authors = ["Émile Grégoire <emile@automatak.com>"]
 edition = "2018"
 
+[lib]
+crate-type = ["cdylib"]
+
 [dependencies]
-oo-bindgen = { path = "../../oo-bindgen" }
-dunce = "1.0"
-heck = "0.3"
 jni = "0.17"
+foo-ffi = { path = "../foo-ffi"}
